// generated/src/app/api-models/agent-address-contact.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class AgentAddressContact {
  seqAgentContact: number;
  seqAgentId: number;
  seqAgentAddress: number;
  contactName: string;
  contactTitle: string;
  phoneNumber: string;
  extension: string;
  faxNumber: string;
  emailAddress: string;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
}